requirements.txt
boto3
aws-cdk-lib==2.68.0